+++ lca/src/main/kotlin/com/dreampany/lca/data/source/dao/NewsDao.kt
@@ -39,8 +39,8 @@ interface NewsDao : BaseDao<News> {
     fun getItemRx(id: Long): Maybe<News>
 
     @Query("select * from news order by publishedOn limit :limit")
-    fun getItems(limit: Long): List<News>
+    fun getItems(limit: Int): List<News>
 
     @Query("select * from news order by publishedOn limit :limit")
-    fun getItemsRx(limit: Long): Maybe<List<News>>
+    fun getItemsRx(limit: Int): Maybe<List<News>>
 }
